pyproject.toml
[project]
name = "DAGTER_HOME"
version = "0.1.0"
description = "A Dagster project with AI and web scraping integrations"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "dagster",
    "google-generativeai",
    "requests",
    "b2sdk",
    "pinecone-client",
    "aiohttp",
    "beautifulsoup4"
]

[project.optional-dependencies]
dev = [
    "dagster-webserver", 
    "pytest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "DAGTER_HOME.scheduler"
code_location_name = "DAGTER_HOME"

[tool.setuptools.packages.find]
exclude=["my_dagster_project_tests"]
